1. Fixed bug with precipitation probability. It is an integer 0-100, not a float between 0-1.0 like originally implemented. 2. Increased ClockScreen and WeatherForecastScreen durations to be 30 seconds each.

+++ data/weather/tomorrowIo.go
@@ -66,8 +66,8 @@ type WeatherTimelinesResponse struct {
 					SunriseTime    string  `json:"sunriseTime"`
 					SunsetTime     string  `json:"sunsetTime"`
 					// Hourly values
-					PrecipitationProbability float64 `json:"precipitationProbability"`
-					UVIndex                  float64 `json:"uvIndex"`
+					PrecipitationProbability int     `json:"precipitationProbability"`
+					UVIndex                  int     `json:"uvIndex"`
 					WindSpeed                float64 `json:"windSpeed"`
 					WindGust                 float64 `json:"windGust"`
 					Temperature              float64 `json:"temperature"`
